requirements.txt
textual>=0.47.0
rich>=13.7.0
beautifulsoup4>=4.12.0
httpx>=0.25.0
jinja2>=3.1.2
qrcode[pil]>=7.4.2
pillow>=10.1.0
